requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.31.0
python-dotenv==1.0.1
openai==1.51.0
pydantic==2.9.2

